refactor: Pong型を新規作成し、WebsocketResponseから関連する型を移動
