+++ packages/ws-signal/src/response/pong.ts
@@ -1,0 +1,5 @@
+/** pingの応答 */
+export type Pong = {
+  action: "pong";
+  message: string;
+};
